Removing mixin classes
