--- src/topotoolbox/gridmixins/__init__.py
@@ -1,6 +1,0 @@
-from .fillsinks import FillsinksMixin
-from .info import InfoMixin
-from .magic import MagicMixin
-from .identifyflats import IdentifyflatsMixin
-
-__all__ = ['FillsinksMixin', 'InfoMixin', 'MagicMixin', 'IdentifyflatsMixin']
